clean up ovl11

+++ src/ft/ftmanager.c
@@ -111,7 +111,7 @@ void ftManagerSetupFileSize(void)
     FTMotionDesc *motion_desc;
 
     rl_setup.table_addr = (uintptr_t)&lLBRelocTableAddr;
-    rl_setup.table_files_num = (uintptr_t)&lLBRelocTableFilesNum;
+    rl_setup.table_files_num = (u32)&lLBRelocTableFilesNum;
     rl_setup.file_heap = NULL;
     rl_setup.file_heap_size = 0;
     rl_setup.status_buffer = NULL;
